one of the method of the hashing
 The file HashInt.txt (in a separate file) contains 100,000 integers all randomly chosen
between 1 and 1,000,000 (there might be some repetitions). This is considered as an array of
integers where the ith row of the file gives you the ith entry of the array. Given below are 9
"target sums", in increasing order: 231552, 234756, 596873, 648219, 726312, 981237,
988331, 1277361, 1283379. You are required to implement the hash table-based
algorithm and determine, for each of the 9 target sums x, whether or not x can be
formed as the sum of two entries in the given array.